+++ CHANGELOG.md
@@ -6,9 +6,17 @@
 
 ## 1.0.1 (29/07/2020)
 
-### Bug fixes
-
+Bug fixes: 
 - Added a margin to the elements of the AuthorTagList.vue component so they would not overlap if the column is too wide
 - Added alphabetical sorting for the lists of publications in the KITOpenAnalytics and KITOpenMetrics pages
 - Fixed the bug, that the metrics page counted all the duplicates
 
+## 1.1.0 (29/07/2020)
+
+Bug fixes:
+- Changed the CSS in MultiSelect.vue so that the labels are now vertically centered
+
+Features added:
+- Added a third selection box to the metrics page, which can be used to select the publications by their POF structure
+    - Also added a new column "Pof Structure" to the list view in the metrics page
+    - Displaying metrics by POF as well